pyproject.toml
[project]
name = "examples"
version = "0.1.0"
description = "Playground repo for iceberg"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.22",
    "duckdb>=1.3.0",
    "jupyter>=1.1.1",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pyiceberg>=0.9.1",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pre-commit-uv>=4.1.4",
    "pyarrow-stubs>=19.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
